packages/py-common/pyproject.toml
[project]
name = "axis-common"
version = "0.1.0"
description = "Shared Python library for Axis services (settings, logging, middleware, errors)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.112.0",
    "pydantic[email]>=2.8.0",
    "pydantic-settings>=2.4.0",
    "asyncpg>=0.30.0",
    "structlog>=24.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "httpx>=0.27.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/axis_common"]

[tool.ruff]
line-length = 100
target-version = "py312"
